+++ AutoFindReplace/Dtos/GeneralOptionsDto.cs
@@ -4,5 +4,7 @@
     {
         public bool UnderSourceControlOnly { get; set; }
         public bool KeepFileOpenAfterSave { get; set; }
+        public bool HideResultsDialogIfNoModifications { get; set; }
+
     }
 }