+++ AutoFindReplace/VSPackage.cs
@@ -68,7 +68,24 @@ namespace AutoFindReplace
                     successMessages.Add(applyChangesMessage);
                 }
 
+                //gregt test this
+                //add to readme.md "features"
+  //              - [x] Publish the extension for VS2017
+//- [x] Auto - size the column widths for Rules in Options
+  //- [x] Added option to suppress results dialog if no modifications applied
+                var showPopUpMessage = false;
                 if (matchingSolutionOpened && (anyRulesProcessed || failureMessages.Count > 0))
+                {
+                    showPopUpMessage = true;
+
+                    if (generalOptionsDto.HideResultsDialogIfNoModifications && changesCount == 0)
+                    {
+                        showPopUpMessage = false;
+                    }
+                }
+
+
+                if (showPopUpMessage)
                 {
                     summaryMessages = messagesHelper.GetSummaryMessages(rulesEnabledForThisSolutionCount, rulesProcesssedSuccessfullyCount, rulesProcesssedUnsuccessfullyCount, changesCount);
                     var userFriendlySuccessMessages = messagesHelper.GetUserFriendlySuccessMessages(successMessages);
@@ -271,6 +288,7 @@ namespace AutoFindReplace
             {
                 UnderSourceControlOnly = generalOptions.UnderSourceControlOnly,
                 KeepFileOpenAfterSave = generalOptions.KeepFileOpenAfterSave,
+                HideResultsDialogIfNoModifications = generalOptions.HideResultsDialogIfNoModifications,
             };
 
             return generalOptionsDto;